Rewrite setup_certs
setup_certs was written to behave like an expect script, it set up a
pseudo terminal and listened for prompts from certutil, when it got a
prompt it recogined it emitted specified input. But this was
fragile. Now instead just write the input to certutil. This also means
we have to use a password file, but we remove the password file when
the process exists.

By default create a new style sql database.

Use argparse for option parsing.

This is a total rewrite.