fix read_data_from_file(), make it accept any file like object
read_data_from_file() was supposed to accept either a string
representing a filename to open or a file object on which it will
call read() to load the contents. However the test for a file object
was too restrictive, it literally checked for a file object which
excluded objects supporting the file interface
(e.g. StringIO). Therefore the test was changed to test if the object
has a read() method.